Move back button outside the country card
