+++ src/components/partners-card/partners-card-styles.module.scss
@@ -1,4 +1,3 @@
-
 @import 'styles/settings';
 @import 'styles/typography-extends';
